src: make attestation check fail in dev devices
Modified factory setup such that secure chip and memory do not have
valid attestation and device keys anymore. This way, when dev
devices are flashed by factory setup and loaded firmware, then the
attestation check that is done while communicating, the computer
will say that the attestation check is failed.

Dev devices can be configured by this method to not to be confused
with prod devices and not to be used for real use cases.